some init cleanup, move visualizer controls
